refactor: split auxiliary_client.py into provider/credential resolution modules
